Small fix
After the realease of version 1.21 I was alerted to a hanging note when importing the music from 
Leisure Suit Larry. Apparently the AGI format requires that a voice is muted when there is no more data for that voice in the file. I have added this.
I have also slightly optimised the drive recognition in the file selector.